+++ app/src/main/java/com/softdesign/devintensive/data/manager/DataManager.java
@@ -1,4 +1,4 @@
-package com.softdesign.devintensive.managers;
+package com.softdesign.devintensive.data.manager;
 
 public class DataManager {
     private static DataManager INSTANCE = null;
@@ -7,14 +7,12 @@ public class DataManager {
     private DataManager() {
         this.mPreferencesManager = new PreferencesManager();
     }
-
     public static DataManager getInstance() {
         if (INSTANCE == null) {
             INSTANCE = new DataManager();
         }
         return INSTANCE;
     }
-
     public PreferencesManager getPreferencesManager() {
         return mPreferencesManager;
     }
